requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
tzdata==2023.3
Werkzeug==2.3.3